benchmarks/ alsa-gtk-bench.cpp whitespace cleanup
This commit fixes layout, whitespace and tabs of the ALSA
GTK benchmark. There are no functional changes - this can
be easily verified by using a source formatting tool, and
applying it to both the original and post-patch version.

A diff -w (ignore whitespace) will show actual code changes
which are only aesthetic (some newline wraps, and blank line
removals)

This patch was done to improve readability, and for me to
get familiar with the code too :) -Harry